requirements.txt
anyio==4.9.0
apify_client==1.9.2
apify_shared==1.3.2
astroid==3.3.9
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
colorama==0.4.6
dill==0.3.9
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isort==6.0.1
mccabe==0.7.0
more-itertools==10.6.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
pycparser==2.22
pylint==3.3.6
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PySocks==1.7.1
pytesseract==0.3.13
python-dotenv==1.1.0
selenium==4.30.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomlkit==0.13.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
urllib3==2.3.0
websocket-client==1.8.0
wget==3.2
wsproto==1.2.0
